Support up to INT_MAX (2^31-1) for the absolute value of css <number> data type.
When converting parsed number from string to double do not use a flot variable to store
the result as it does not have the necessary precision to represent exactly all the values
in the 32-bit integer range. Use a double instead.

bug 335346

(forwardport kdelibs commit 6197b21)

+++ src/css/parser.h
@@ -1,19 +1,19 @@
 /* A Bison parser, made by GNU Bison 2.5.1.  */
 
 /* Bison interface for Yacc-like parsers in C
-
+   
       Copyright (C) 1984, 1989-1990, 2000-2012 Free Software Foundation, Inc.
-
+   
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.
-
+   
    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU General Public License for more details.
-
+   
    You should have received a copy of the GNU General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.  */
 
@@ -26,75 +26,80 @@
    special exception, which will cause the skeleton and the resulting
    Bison output files to be licensed under the GNU General Public
    License without this special exception.
-
+   
    This special exception was added by the Free Software Foundation in
    version 2.2 of Bison.  */
+
 
 /* Tokens.  */
 #ifndef YYTOKENTYPE
 # define YYTOKENTYPE
-/* Put the tokens into the symbol table, so that GDB and other debuggers
-   know about them.  */
-enum yytokentype {
-    REDUCE = 258,
-    S = 259,
-    SGML_CD = 260,
-    INCLUDES = 261,
-    DASHMATCH = 262,
-    BEGINSWITH = 263,
-    ENDSWITH = 264,
-    CONTAINS = 265,
-    STRING = 266,
-    IDENT = 267,
-    NTH = 268,
-    HASH = 269,
-    HEXCOLOR = 270,
-    IMPORT_SYM = 271,
-    PAGE_SYM = 272,
-    MEDIA_SYM = 273,
-    FONT_FACE_SYM = 274,
-    CHARSET_SYM = 275,
-    NAMESPACE_SYM = 276,
-    KHTML_RULE_SYM = 277,
-    KHTML_DECLS_SYM = 278,
-    KHTML_VALUE_SYM = 279,
-    KHTML_MEDIAQUERY_SYM = 280,
-    KHTML_SELECTORS_SYM = 281,
-    IMPORTANT_SYM = 282,
-    MEDIA_ONLY = 283,
-    MEDIA_NOT = 284,
-    MEDIA_AND = 285,
-    QEMS = 286,
-    EMS = 287,
-    EXS = 288,
-    PXS = 289,
-    CMS = 290,
-    MMS = 291,
-    INS = 292,
-    PTS = 293,
-    PCS = 294,
-    DEGS = 295,
-    RADS = 296,
-    GRADS = 297,
-    MSECS = 298,
-    SECS = 299,
-    HERZ = 300,
-    KHERZ = 301,
-    DPI = 302,
-    DPCM = 303,
-    DIMEN = 304,
-    PERCENTAGE = 305,
-    FLOAT = 306,
-    INTEGER = 307,
-    URI = 308,
-    FUNCTION = 309,
-    NOTFUNCTION = 310,
-    UNICODERANGE = 311
-};
+   /* Put the tokens into the symbol table, so that GDB and other debuggers
+      know about them.  */
+   enum yytokentype {
+     REDUCE = 258,
+     S = 259,
+     SGML_CD = 260,
+     INCLUDES = 261,
+     DASHMATCH = 262,
+     BEGINSWITH = 263,
+     ENDSWITH = 264,
+     CONTAINS = 265,
+     STRING = 266,
+     IDENT = 267,
+     NTH = 268,
+     HASH = 269,
+     HEXCOLOR = 270,
+     IMPORT_SYM = 271,
+     PAGE_SYM = 272,
+     MEDIA_SYM = 273,
+     FONT_FACE_SYM = 274,
+     CHARSET_SYM = 275,
+     NAMESPACE_SYM = 276,
+     KHTML_RULE_SYM = 277,
+     KHTML_DECLS_SYM = 278,
+     KHTML_VALUE_SYM = 279,
+     KHTML_MEDIAQUERY_SYM = 280,
+     KHTML_SELECTORS_SYM = 281,
+     IMPORTANT_SYM = 282,
+     MEDIA_ONLY = 283,
+     MEDIA_NOT = 284,
+     MEDIA_AND = 285,
+     QEMS = 286,
+     EMS = 287,
+     EXS = 288,
+     PXS = 289,
+     CMS = 290,
+     MMS = 291,
+     INS = 292,
+     PTS = 293,
+     PCS = 294,
+     DEGS = 295,
+     RADS = 296,
+     GRADS = 297,
+     MSECS = 298,
+     SECS = 299,
+     HERZ = 300,
+     KHERZ = 301,
+     DPI = 302,
+     DPCM = 303,
+     DIMEN = 304,
+     PERCENTAGE = 305,
+     FLOAT = 306,
+     INTEGER = 307,
+     URI = 308,
+     FUNCTION = 309,
+     NOTFUNCTION = 310,
+     UNICODERANGE = 311
+   };
 #endif
 
+
+
 #if ! defined YYSTYPE && ! defined YYSTYPE_IS_DECLARED
-typedef union YYSTYPE {
+typedef union YYSTYPE
+{
+
 
     CSSRuleImpl *rule;
     CSSSelector *selector;
@@ -104,7 +109,7 @@ typedef union YYSTYPE {
     CSSMediaRuleImpl *mediaRule;
     CSSRuleListImpl *ruleList;
     ParseString string;
-    float val;
+    double val;
     int prop_id;
     unsigned int attribute;
     unsigned int element;
@@ -120,9 +125,14 @@ typedef union YYSTYPE {
     QList<khtml::MediaQueryExp *> *mediaQueryExpList;
     khtml::MediaQuery::Restrictor mediaQueryRestrictor;
 
+
+
 } YYSTYPE;
 # define YYSTYPE_IS_TRIVIAL 1
 # define yystype YYSTYPE /* obsolescent; will be withdrawn */
 # define YYSTYPE_IS_DECLARED 1
 #endif
+
+
+
